spotify_etl_pipeline/requirements.txt
spotipy>=2.24.0
pandas>=2.2.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
matplotlib>=3.9.0
plotly>=5.22.0
numpy>=1.26.4
requests>=2.32.0

